requirements.txt
pillow
numba
cudatoolkit